backend/requirements.txt
# for fastapi server
fastapi==0.110.0
uvicorn==0.27.1

# array and image processing
numpy==1.26.4
pillow==10.3.0

# pydantic for schema validation
pydantic==2.7.1
python-multipart==0.0.9

# machine learning models
scikit-learn==1.5.0
torch==2.2.2
